+++ examples/react-router-paneling-framework/app/routes/paneling.tsx
@@ -38,7 +38,7 @@ export async function clientLoader(args: Route.ClientLoaderArgs) {
         ),
         indexComponent: IndexPage,
         max: 8,
-        separator: ';',
+        extrasSeparator: ';',
         panels: {
             info: createCustomPanel(
               InfoContent,