+++ examples/react-router-paneling-zustand/app/components/panels/CustomPanel.tsx
@@ -1,3 +1,4 @@
+// client-only component (uses Zustand)
 "use client";
 
 import { type MouseEventHandler, type RefObject, type TouchEventHandler, useCallback, useEffect, useRef } from 'react'