+++ examples/react-router-paneling-zustand/app/routes/paneling.tsx
@@ -39,7 +39,7 @@ export async function clientLoader(args: Route.ClientLoaderArgs) {
         ),
         indexComponent: IndexPage,
         max: 8,
-        separator: ';',
+        extrasSeparator: ';',
         panels: {
             info: createCustomPanel(
               InfoContent,